requirements.txt
reflex>=0.8.14,<0.9
pyserial>=3.5,<4
